+++ src/ui_event/dialog_ui.py
@@ -226,7 +226,7 @@ class Dialog(QDialog):
         :return:
         """
         # 在这里你可以添加任何你需要在对话框关闭时执行的代码
-        logger.debug('config msg dialog Dialog is cancel closing!')
+        logger.debug('log show  dialog  is cancel closing!')
         constants.edit_config_msg_visible = False
         # 调用基类的 closeEvent 方法以确保对话框正常关闭
         super(Dialog, self).reject()
@@ -238,7 +238,7 @@ class Dialog(QDialog):
         :return:
         """
         # 在这里你可以添加任何你需要在对话框关闭时执行的代码
-        logger.debug('config msg dialog Dialog is close closing!')
+        logger.debug('log show  dialog is close closing!')
         constants.edit_config_msg_visible = False
         # 调用基类的 closeEvent 方法以确保对话框正常关闭
         super(Dialog, self).closeEvent(event)